\name{add.examples.to.package}
\alias{add.examples.to.package}
\title{Add examples to Rd files of a package}
\usage{
  add.examples.to.package(package,
    package.path = path.package(package),
    example.path = NULL, example.files = NULL)
}
\description{
  Add examples to Rd files of a package
}

